Create Scheme Eligibility Test Documents
The main objective of this project is to know the government schemes that you are eligible to apply for. In this project the user can enter his details and see which scheme he is eligible for.
In this project I created a Test Scenarios ,Test Cases and Bug report for this project.